fix(wma,hma,linearreg): compute period-scaled dividers in double (#142)
Several period-scaled integer expressions were evaluated in int32 before
being stored to a double, overflowing (signed UB) well inside the advertised
[.,100000] period range:

- WMA/HMA  divider = (n*(n+1))>>1  -- the int product overflows at n>=46341
  (the issue's 65536 is only where the post-shift *value* overflows; the
  multiply happens first).
- LINEARREG / LINEARREG_SLOPE / _INTERCEPT / _ANGLE / TSF
  SumXSqr = n*(n-1)*(2n-1)/6  -- cubic in int32, overflows at n>=1025 (the
  binding site), plus SumX = n*(n-1)*0.5 at n>=46342.

Fix widens the operands to double at the source (ta_codegen/input); the WMA/HMA
divider locals become double (used only as periodSum/divider). Bit-identical
for every currently-valid period -- triangular/pyramidal numbers <=2^53 are
exactly representable, so xlang-hash and fuzz-064 both stay at zero diffs --
and mathematically exact for the previously-overflowing periods. The max period
stays 100000: no cap fixes this (the cubic overflows at 1025, a cap would have
to be <=1024) and the unguarded variants skip the range check anyway. Also
converges a latent C<->Rust divergence -- Rust already computed the WMA divider
in 64-bit, so above the int32 threshold C wrapped while Rust was correct.

Regenerated C (src/ta_func: guarded + unguarded + TA_S_ stream), Rust (drops
the now-dead `as f64` divide-site cast; stream state field int->f64) and Java
(Core.java + server). No generator change: stream state field types are
type-driven from the batch-local decl. .NET P/Invokes C.

Tests (both sabotage-proven):
- backend_suite.rs: assert all 7 sites render as double across C/Rust/Java, no
  int32 `>> 1` / `/ 6` -- source-form regression guard at generation time.
- test_period_boundary.c: run the LR family at period 1025 on a ~16 KB linear
  ramp (exact least-squares fit => closed-form output). The teeth are the
  --sanitize (UBSan) nightly, which traps the int32 product; a plain -O3
  release build folds the overflow to the correct value (the optimizer is
  entitled to assume no signed overflow), so the value check there is
  coherence-only.

+++ src/ta_func/ta_LINEARREG_ANGLE.c
@@ -141,8 +141,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_ANGLE( int    startIdx,
    /* Index into the output. */
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    /* Prime the two data-dependent window sums for the first output with a
     * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -220,8 +220,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_ANGLE_Unguarded( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -299,8 +299,8 @@ TA_RetCode TA_S_LINEARREG_ANGLE( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -364,8 +364,8 @@ TA_RetCode TA_S_LINEARREG_ANGLE_Unguarded( int    startIdx,
    outIdx = 0;
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
-   SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-   SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+   SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+   SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
    SumXY = 0;
    SumY = 0;
@@ -510,8 +510,8 @@ TA_RetCode TA_LINEARREG_ANGLE_OpenInternal( struct TA_LINEARREG_ANGLE_Stream **s
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
@@ -653,8 +653,8 @@ TA_LIB_API TA_RetCode TA_LINEARREG_ANGLE_OpenAndFill( TA_LINEARREG_ANGLE_Stream 
       /* Index into the output. */
       today = startIdx;
       trailingIdx = startIdx - lookbackTotal;
-      SumX = optInTimePeriod * (optInTimePeriod - 1) * 0.5;
-      SumXSqr = optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6;
+      SumX = (double)optInTimePeriod * (optInTimePeriod - 1) * 0.5;
+      SumXSqr = (double)optInTimePeriod * (optInTimePeriod - 1) * (2 * optInTimePeriod - 1) / 6.0;
       Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
       /* Prime the two data-dependent window sums for the first output with a
        * one-time full-window scan. SumX/SumXSqr/Divisor are period-only constants;
