+++ ta_codegen/input/linearreg_intercept/linearreg_intercept.c
@@ -72,8 +72,8 @@ TA_RetCode linearreg_intercept(int startIdx, int endIdx,
    today = startIdx;
    trailingIdx = startIdx - lookbackTotal;
 
-   SumX = optInTimePeriod * ( optInTimePeriod - 1 ) * 0.5;
-   SumXSqr = optInTimePeriod * ( optInTimePeriod - 1 ) * ( 2 * optInTimePeriod - 1 ) / 6;
+   SumX = (double)optInTimePeriod * ( optInTimePeriod - 1 ) * 0.5;
+   SumXSqr = (double)optInTimePeriod * ( optInTimePeriod - 1 ) * ( 2 * optInTimePeriod - 1 ) / 6.0;
    Divisor = SumX * SumX - optInTimePeriod * SumXSqr;
 
    /* Prime the two data-dependent window sums for the first output with a
